BUGFIX: Mouse capture not released on wxAuiManager
Should a pane resize fail on DoEndResizeAction the mouse was recaptured while discarding the operation to perform.
This caused a host of UI problems as the mouse was not released even after LeftUp and eventually an app crash as
a posterior LeftDown over a sizer would try to recapture the mouse, failing on an assert.
This bug could be triggered by trying to resize a pane with no non-fixed one to the right on the dock,
i.e. a left dock with two resizable panes and a fixed one, from top to bottom, trying to move the last sizer.