Fix key codes in key up/down events with non-US keyboard layout
Changes using XKB for getting key codes for the events in wxGTK resulted
in always using the key codes for the US keyboard which was incompatible
and arguably wrong too.

Change the logic to only use XKB when we don't have a valid key code in
the current layout to restore the old behaviour.

Closes #23680.

Closes #23818.