Handle all unknown error codes as wxWEBVIEW_NAV_ERR_OTHER
This avoids many warnings about unhandled enum values in the switch
statement.

It would be better to map all the error code to the correct category,
but it seems like this enum is being modified relatively often in
Chromium/CEF, so for now just use wxWEBVIEW_NAV_ERR_OTHER for all of the
unknown values.

Also handle ERR_NONE here, even if it's not clear if we can ever get it.