Stop using CefBrowser::LoadString()
This function doesn't exist in the latest CEF versions and apparently
didn't work for a long time before it, see chromiumembedded/cef#2586, so
stop using it and use data: URI instead for loading the fixed text.

This replacement is not quite perfect, notably it doesn't show the
correct URI in the title bar, but better than nothing.