Construct temporary CefRefPtr instead of using a cast
Even if CefRefPtr<T> has the same representation as T*, it is still
wrong to cast the latter to the former.

Instead, construct a temporary smart pointer object: this will make
extra calls to AddRef() and Release() but it's really not a big deal
when creating a new browser object.