Simplify ClientHandler construction
No real changes, just pass the associated wxWebView object to the ctor
instead of using a separate function for this, this is shorter and makes
it more clear that this pointer is required and also allows to make it a
reference rather than a (never null) pointer.

Also remove unnecessary dtor and initialize members in their declaration
when possible.