Remove wxInitData::argvOrig which is not needed, finally
We never really used the "original" version of the command line, i.e.
before it was (possibly) modified by gtk_init_check(), so don't bother
keeping it and just free the pointers corresponding to the arguments
consumed by GTK in wxGTK code.

There should be no real changes, this is just a simplification.