[EGL] Use eglGetProcAddress() to get EGLDisplay
This modifies the EGLDisplay initialization to pick the appropriate
function at runtime, depending on what is available. The preferred
function is `eglGetPlatformDisplay()`, which is only available on EGL
1.5+. As a fallback, `eglGetPlatformDisplayEXT()` is used. This is
available on some platforms using EGL 1.4, but is no longer available on
EGL 1.5. If all else fails, use the older `eglGetDisplay()` API.