Don´t search for wxstd MO files with version number, just look for wxstd.mo
Searching for translated a wxstd-3.3.mo file is supposed to fallback to searching for a translated wxstd.mo file then. However, it does not work that way because English is the default wxLanguage argument that AddStdCatalog passes to AddCatalog(). What happens is that AddCatalog can't find a translated wxstd-3.3.mo file and simply falls back to English and returns true. Then AddStdCatalog accepts that returned true value and never attempts to call AddCatalog again with just "wxstd". Hence, my "fix" is to simply look for a translated wxstd.mo.