Added QtBlendMaskWithAlpha() to wxBitmap under wxQt
Blend mask with alpha channel and remove the mask.

We need this function for performance reasons. e.g. the wxImageList (used by
wxTreeCtrl and wxListCtrl) calls this function to convert the mask (if any)
to alpha. Because the controls (under wxQt) using the image list ignore the
mask on the bitmaps and expect alpha instead.

This is similar to MSWBlendMaskWithAlpha() found in wxMSW port.