Optimize wxInputStream::SeekI() from start or current position
It's not clear whether it's worth explicitly checking for no-op seeks at
all, but unnecessarily calling GetLength() when doing it is certainly a
pessimization rather than optimization, so don't do it.

And don't bother checking for no-op seeks when seeking from end at all,
as it's not obviously faster than just performing the seek and the
existing code didn't work correctly anyhow, as it compared the current
position with "size-pos" instead of the correct "size+pos", so removing
it certainly shouldn't make things worse.

Closes #23895.