Disable blocking in eglSwapBuffers() when using X11 too
Unfortunately we can't let EGL under XWayland block neither, even if we
can't update the window efficiently in this case, as we could do when
using real Wayland in 194a7be33f (Improve wxGLCanvasEGL refresh logic
under Wayland, 2023-05-17), because otherwise occluding any window with
a wxGLCanvasEGL inside it slows down the entire program to a crawl as
all events are generated at a rate of 1 per second only.

Choose the lesser of two evils and update even completely hidden
windows, which is inefficient, but at least allows to also update the
other ones, unlike before.

Also note that we must call eglSwapInterval() in SwapBuffers() and that
calling it in CreateSurface() somehow has no effect when using Sway WM,
even though it would also solve the problem under GNOME/mutter.

Closes #23512.

Co-authored-by: Tristan Daniel <tdaniel22@users.noreply.github.com>