Fix the wxStyledTextCtrl autocompletion popup
The changes implemented on the wxPopupWindow have broken the autocompletion popup of this control.
The fix simply changes the mode of operation of the popup window, avoiding the breakage. I also
don't see why the popup window was created with a wxPU_CONTAINS_CONTROLS style flag since the
list control does not accept focus.