Fix a crash in wxAUI
wxAUI has a special window "wxTabFrame" for which GetHandle() returns nullptr.
This is a problem because this commit 3c68696 (Make wxTopLevelWindow::SetSizeHints()
work in wxQt) doesn't account for the rare situation in which a wxWindow is actually
created without the underlying handle, resulting in a crash when we try to call
Set{Min,Max}Size() for this type of windows.