Fix initial size of wxPreferencesEditor in wxGTK under Wayland
Setting full size of the dialog before showing it doesn't work, but
setting its client size does.

It could be better to change wxWindow::Fit() itself to do this, but this
might have other effects, while this change should be pretty safe, i.e.
not break anything for the other platforms while fixing a big usability
problem with wxGTK/Wayland: as this dialog is not resizeable, making it
too small initially, as happened before, meant that parts of it couldn't
be used at all.