CMake: fix RegularExpression::compile errors
When the wxWidgets sources are in a path containing '++', CMake would show errors on the command line:
[cmake] RegularExpression::compile(): Nested *?+.
[cmake] RegularExpression::compile(): Error in compile

This happens because the full path was interpreted as a regular expression when creating the source_groups.
Without wxSOURCE_DIR the regular expressions still match and create correct source groups, so remove it.

Fixes #22738