Fixing MSW crash with empty wxListCtrl in dark mode
Changes of 9710592cc4 (Erase unwanted list view separators in dark mode
under Windows 11, 2023-06-05) wrongly kept the check for "itemCount",
verifying that the control is not empty, only in the "needToDraw" case,
but it is needed in the "needToErase" case as well because it also uses
the bottom item index which is invalid if there are no items.

Fix this by always checking that the control is non-empty before doing
anything else.

Closes #23960.

Co-authored-by: Vadim Zeitlin <vadim@wxwidgets.org>