Add support for initializer_list to wx dynamic arrays
While these array classes are deprecated in the user code, they're still
used, for compatibility, in many places in wxWidgets API and allowing to
create them from initializer_list makes using it more ergonomic as it's
now possible to just pass an initializer list of items to fill the
control with, for example, instead of appending them one by one.

This is the equivalent of 4d62df488b (Add support for initializer_list
to wx dynamic arrays, 2023-03-02) from master.

See #23309.

Closes #23966.