Fix refreshing wxGenericListCtrl with multiple selection
Changing the selection could behave wrongly and result in an assert
failure as the selection anchor could have become invalid, due to the
change of the number of items in the control.

Fix this by invalidating it when this happens.

(cherry picked from commit 5885eea5ea72434311f2c706f57cff92a09ade91)