Fix losing the other selection contents in wxClipboard
The memory leak fix in b52728a62a (Fix memory leak of wxClipboard data
on exit, 2023-06-21) was too eager and destroyed not just the currently
used selection (e.g. primary), but also the other one (secondary) when
Clear() was called, which was wrong and resulted in the loss of
clipboard contents when discarding the primary selection in wxSTC, for
example.

Fix this by only deleting the object corresponding to the selection
which we're clearing.

See #23661, #23988.

(cherry picked from commit dc6a0c069bdde714f22cb0459efd9ad186ce7179)